Working game

+++ Assets/Scripts/User System/UserData.cs
@@ -4,7 +4,7 @@ public class UserData : ScriptableObject
 {
 	public string userName;
 	public string userNameColor;
-	public long points;
+	public int points;
 	public bool isSubscriber;
 	public bool isTurbo;
 	public bool isMod;
